Split provider types into more manageable chunks
