+++ src/app/screens/Data/types.tsx
@@ -1,4 +1,4 @@
-import { ProvidedDataTypes, ProviderDatum } from 'main/providers/types';
+import { ProvidedDataTypes, ProviderDatum } from "main/providers/types/Data";
 
 export type GroupedData =  { [key: string]: ProviderDatum<string, ProvidedDataTypes>[] };
 export type DeletedData = { [key: string]: number[] };
