+++ src/main/lib/repository/utilities/generate-parsed-commit.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import { getParserByFileName } from 'main/providers/parsers';
 import parseSchema from './parse-schema';
-import { ProviderDatum } from 'main/providers/types';
+import { ProviderDatum } from "main/providers/types/Data";
 import { Blob, TreeEntry } from 'nodegit';
 import parseCsv from './parse-csv';
 import parseOpenDataRights, { OpenDataRightsDatum } from './parse-open-data-rights';
